+++ controller/src/controller.rs
@@ -20,10 +20,9 @@ use crate::core::core;
 use crate::core::core::Transaction;
 use crate::impls::{FileWalletCommAdapter, HTTPWalletCommAdapter, KeybaseWalletCommAdapter};
 use crate::keychain::Keychain;
-use crate::libwallet::slate::Slate;
+use crate::libwallet::{Slate, CbData, InitTxArgs, OutputCommitMapping, SendTXArgs};
 use crate::libwallet::types::{
-	CbData, InitTxArgs, NodeClient, OutputCommitMapping, SendTXArgs, TxLogEntry, WalletBackend,
-	WalletInfo,
+	NodeClient, TxLogEntry, WalletBackend, WalletInfo,
 };
 use crate::libwallet::{Error, ErrorKind};
 use crate::util::to_base64;
