+++ controller/tests/check.rs
@@ -28,7 +28,8 @@ use self::keychain::ExtKeychain;
 use grin_wallet_libwallet as libwallet;
 use impls::test_framework::{self, LocalWalletClient, WalletProxy};
 use impls::FileWalletCommAdapter;
-use libwallet::types::{InitTxArgs, WalletInst};
+use libwallet::InitTxArgs;
+use libwallet::types::WalletInst;
 use std::fs;
 use std::thread;
 use std::time::Duration;
