+++ controller/tests/repost.rs
@@ -25,8 +25,7 @@ use grin_wallet_util::grin_util as util;
 use self::core::global;
 use self::core::global::ChainTypes;
 use self::keychain::ExtKeychain;
-use self::libwallet::slate::Slate;
-use self::libwallet::types::InitTxArgs;
+use self::libwallet::{Slate, InitTxArgs};
 use impls::test_framework::{self, LocalWalletClient, WalletProxy};
 use impls::FileWalletCommAdapter;
 use std::fs;
