+++ controller/tests/restore.rs
@@ -26,7 +26,8 @@ use self::core::global;
 use self::core::global::ChainTypes;
 use self::keychain::{ExtKeychain, Identifier, Keychain};
 use self::libwallet::slate::Slate;
-use self::libwallet::types::{AcctPathMapping, InitTxArgs};
+use self::libwallet::InitTxArgs;
+use self::libwallet::types::AcctPathMapping;
 use impls::test_framework::{self, LocalWalletClient, WalletProxy};
 use std::fs;
 use std::sync::atomic::Ordering;
