+++ controller/tests/self_send.rs
@@ -26,7 +26,7 @@ use self::core::global::ChainTypes;
 use self::keychain::ExtKeychain;
 use grin_wallet_libwallet as libwallet;
 use impls::test_framework::{self, LocalWalletClient, WalletProxy};
-use libwallet::types::InitTxArgs;
+use libwallet::InitTxArgs;
 use std::fs;
 use std::thread;
 use std::time::Duration;
