+++ controller/tests/transaction.rs
@@ -26,8 +26,8 @@ use self::core::core::transaction;
 use self::core::global;
 use self::core::global::ChainTypes;
 use self::keychain::ExtKeychain;
-use self::libwallet::slate::Slate;
-use self::libwallet::types::{InitTxArgs, OutputStatus};
+use self::libwallet::{InitTxArgs, Slate};
+use self::libwallet::types::OutputStatus;
 use impls::test_framework::{self, LocalWalletClient, WalletProxy};
 use std::fs;
 use std::thread;
