+++ libwallet/src/api_impl.rs
@@ -23,3 +23,4 @@
 
 pub mod foreign;
 pub mod owner;
+pub mod types;
