+++ libwallet/src/lib.rs
@@ -50,3 +50,4 @@ pub use crate::error::{Error, ErrorKind};
 pub use crate::slate::Slate;
 pub use crate::slate_versions::{SlateVersion, VersionedSlate};
 pub use internal::restore::{check_repair, restore};
+pub use api_impl::types::{SendTXArgs, InitTxArgs, InitTxSendArgs, BlockFees, CbData, OutputCommitMapping, NodeHeightResult};
